# src/docs/swagger.yaml
basePath: /
definitions:
  github_com_gothi_vouchrs_src_delivery_http_response.APIError:
    properties:
      code:
        example: NOT_FOUND
        type: string
      message:
        example: listing not found
        type: string
    type: object
  github_com_gothi_vouchrs_src_delivery_http_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.APIError'
      meta: {}
      success:
        example: true
        type: boolean
    type: object
...
  github_com_gothi_vouchrs_src_internal_domain_entity.Listing:
    properties:
      brandID:
        type: string
      buyerPrice:
        description: what buyer pays
        type: number
      codeEncrypted:
        description: AES-256-GCM encrypted
        type: string
      codeHash:
        description: SHA-256 for duplicate check
        type: string
      createdAt:
        type: string
      discountPct:
        description: e.g. 9.0 for 9% off
        type: number
      faceValue:
        description: e.g. 1000.00
        type: number
      gate1At:
        type: string
      id:
        type: string
      isPool:
        description: true = CardSwap pool
        type: boolean
      lockBuyerID:
        type: string
      lockExpiresAt:
        type: string
      sellerID:
        type: string
      sellerPayout:
        description: what seller receives
        type: number
      soldAt:
        type: string
      status:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.ListingStatus'
      updatedAt:
        type: string
      verifiedBalance:
        description: balance at Gate 1
        type: number
    type: object
  github_com_gothi_vouchrs_src_internal_domain_entity.ListingStatus:
    enum:
    - LIVE
    - LOCKED
    - SOLD
    - EXPIRED
    - CANCELLED
    - FRAUD_HOLD
    type: string
    x-enum-varnames:
    - ListingStatusLive
    - ListingStatusLocked
    - ListingStatusSold
    - ListingStatusExpired
    - ListingStatusCancelled
    - ListingStatusFraudHold
  github_com_gothi_vouchrs_src_internal_domain_entity.PoolGroup:
    properties:
      activeCount:
        description: LIVE listings in pool
        type: integer
      avgSellTimeMins:
        type: number
      brandID:
        type: string
      buyerPrice:
        type: number
      createdAt:
        type: string
      discountPct:
        type: number
      faceValue:
        type: number
      id:
        type: string
      recommendedPrice:
        description: seller receives this
        type: number
      updatedAt:
        type: string
    type: object
...
  github_com_gothi_vouchrs_src_internal_domain_port.MarketplaceResult:
    properties:
      individualListings:
        items:
          $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.Listing'
        type: array
      poolGroups:
        items:
          $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.PoolGroup'
        type: array
      total:
        type: integer
    type: object
...
  src_delivery_http_handler.createListingBody:
    properties:
      accept_pool:
        example: true
        type: boolean
      brand_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      card_code:
        example: AMZN-XXXX-XXXX-XXXX
        minLength: 4
        type: string
      custom_discount:
        example: 5
        type: number
      face_value:
        example: 1000
        type: number
    required:
    - brand_id
    - card_code
    - face_value
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cardswap.in
    name: CardSwap Support
  description: P2P gift card exchange marketplace. Card codes are AES-256 encrypted
    at rest and **never returned in API responses** — they are delivered to the buyer's
    registered email only.
  title: CardSwap India API
  version: "1.0"
paths:
...
  /api/v1/listings:
    post:
      consumes:
      - application/json
      description: Post a gift card for sale. Runs Gate 1 Qwikcilver verification.
        The card code is AES-256 encrypted at rest and never returned by any API endpoint
        — it is delivered to the buyer's email only.
      parameters:
      - description: Listing details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/src_delivery_http_handler.createListingBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.Listing'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "409":
          description: Duplicate card code (CONFLICT)
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "422":
          description: Gate 1 failed (VERIFICATION_FAILED) or daily limit (LISTING_LIMIT_REACHED)
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: Create a listing (seller)
      tags:
      - listings
  /api/v1/listings/{id}:
    delete:
      description: Cancel a LIVE listing. Only the original seller can cancel. Returns
        403 if called by any other user.
      parameters:
      - description: Listing UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: listing cancelled
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "403":
          description: Not the listing owner
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: Cancel a listing (seller)
      tags:
      - listings
    get:
      description: Returns listing details. The code_encrypted and code_hash fields
        are always stripped.
      parameters:
      - description: Listing UUID
        example: '"550e8400-e29b-41d4-a716-446655440000"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.Listing'
              type: object
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      summary: Get a listing by ID
      tags:
      - listings
...
  /api/v1/marketplace:
    get:
      description: Returns CardSwap pool groups (aggregated inventory) at the top
        and individual seller listings below. Pool groups show the best available
        price for each brand+denomination combination.
      parameters:
      - description: Filter by brand UUID
        in: query
        name: brand_id
        type: string
      - description: Sort field
        enum:
        - price_asc
        - price_desc
        - discount_desc
        in: query
        name: sort_by
        type: string
      - description: Page number (default 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_port.MarketplaceResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      summary: Browse marketplace
      tags:
      - marketplace
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"